pyproject.toml
[project]
name = "github-star-mcp"
version = "0.1.0"
description = "GitHub Stars MCP Server with RAG and Gitea backup"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.1.0",
    "lancedb>=0.12.0",
    "sqlmodel>=0.0.14",
    "httpx>=0.27.0",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.0",
    "gitpython>=3.1.0",
    "sentence-transformers>=2.3.0",
    "anthropic>=0.25.0",
    "uvicorn>=0.27.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-ollama>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"

[project.scripts]
github-star-mcp = "github_star_mcp.main:main"
